interface Generator changed to abstract class AbstractGenerator
the motivation to this unusual decision is that I can force implementors
implement *protected* method instead of public one. Because of that
implementors(descendants) will create that method, but callers won't be
able to call it.

Benefit of that is improved builder suggestion,
as generator has to have method create to trigger the creation, but when
writing the builder code it might be written at places, where there is
no point of writing that method. With this construct and trick with
class Initialize — which is in same package, thus this method can call it,
we get code which reads better, and it's harder to
call create at places where it shouldn't be used.