rename InitializationContext to InitializationConfiguration
I think this makes it more understandable what this class is intended for.
Along with that I had to rename ConfigurationBuilder to
InitializationConfigurationBuilder. ConfigurationBuilder was created
at moment when I already assumed that InitializationContext ins't good
name. So after this, both should by named in sync. I'm not sure if
future won't show, that there isn't any other configuration needed,
and then we might rename it further to just Configuration respective
ConfigurationBuilder.