removed "duplicity" among overloaded AbstractGenerator create methods
AbstractGenerator declared 2 create methods. First for creating single
instance, and second for create list of such instances. This was intended
to simplify stuff, but the same is possible if one just wraps whichever
generator is passed to deleted method into Generators.list() generator.
Doing so will also make more options to be available:
generating set/stream ..., specifying variable sizes, specifying
requested collection implementation etc. So the deleted method did not
add anything which wasn't doable otherwise, did not offer anything extra,
and it was just creating a lot of code to be maintained.